Update classification tests for is_basic
